requirements.txt
asgiref==3.8.1
Django==5.1.1
django-unfold==0.40.0
sqlparse==0.5.1
tzdata==2024.2
whitenoise==6.8.2
